Checkpointing state for refactor
