+++ network/src/lib.rs
@@ -21,8 +21,6 @@
 #![cfg_attr(nightly, feature(doc_cfg, external_doc))]
 #![cfg_attr(nightly, doc(include = "../documentation/concepts/network_server.md"))]
 
-// #[macro_use]
-// extern crate lazy_static;
 #[macro_use]
 extern crate log;
 #[macro_use]
@@ -36,6 +34,9 @@ pub mod internal;
 
 pub mod request_manager;
 pub use request_manager::*;
+
+pub mod response_manager;
+pub use response_manager::*;
 
 pub mod server;
 pub use server::*;
